Move the configuration code into its own package.
