+++ src/main/java/polyglot/config/BUILD
@@ -1,0 +1,15 @@
+package(default_visibility = ["//visibility:public"])
+
+java_library(
+  name = "config",
+  srcs = glob(['*.java']),
+  deps = [
+    "//src/main/proto:config_proto",
+    "//src/main/java/polyglot/protobuf",
+    "//third_party/args4j",
+    "//third_party/grpc",
+    "//third_party/guava",
+    "//third_party/netty",
+    "//third_party/protobuf",
+  ],
+)
