+++ src/main/java/polyglot/config/ConfigurationLoader.java
@@ -1,4 +1,4 @@
-package polyglot;
+package polyglot.config;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -6,8 +6,8 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Optional;
 
-import com.google.api.client.util.Joiner;
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
 import com.google.protobuf.util.JsonFormat;
