+++ src/test/java/polyglot/grpc/CompositeStreamObserverTest.java
@@ -10,6 +10,7 @@ import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
 import io.grpc.stub.StreamObserver;
+import polyglot.grpc.CompositeStreamObserver;
 
 
 /** Unit tests for {@link CompositeStreamObserver}. */
