+++ src/test/java/polyglot/protobuf/ProtoMethodNameTest.java
@@ -4,6 +4,8 @@ import static com.google.common.truth.Truth.assertThat;
 
 import org.junit.Test;
 
+import polyglot.protobuf.ProtoMethodName;
+
 /** Unit tests for {@link ProtoMethodName}. */
 public class ProtoMethodNameTest {
   @Test
