+++ src/test/java/polyglot/protobuf/ServiceResolverTest.java
@@ -5,6 +5,8 @@ import org.junit.Test;
 
 import com.google.protobuf.DescriptorProtos.FileDescriptorSet;
 
+import polyglot.protobuf.ProtoMethodName;
+import polyglot.protobuf.ServiceResolver;
 import polyglot.test.TestProto;
 import polyglot.test.foo.FooProto;
